src/requirements.txt
pandas
kedro==0.17.1
duckdb
polars
ibis-framework[duckdb,mssql]

# pin bugged dynaconf
dynaconf==3.1.5

# kedro autogen
black==v19.10b0
flake8>=3.7.9, <4.0
ipython==7.10
isort>=4.3.21, <5.0
jupyter~=1.0
jupyter_client>=5.1, <7.0
jupyterlab==0.31.1
kedro==0.17.1
nbstripout==0.3.3
pytest-cov~=2.5
pytest-mock>=1.7.1, <2.0
pytest~=6.1.2
wheel==0.32.2
